# openapi_complete_v5.3.yaml
openapi: 3.1.0
info:
  title: DentalKiosk Shopware E-Commerce API
  description: 'Complete API for the DentalKiosk chatbot middleware used with Shopware
    6 and Dify.


    Features:

    - Product search & details

    - Cart action queue (add/remove/update/clear)

    - Orders lookup and reorder (by orderNumber)

    - Multimodal pipeline (image -> extract -> search -> top3 -> confirm)

    - Analytics ingestion + summary

    '
  version: 5.3.0
  contact:
    name: DentalKiosk Support
    url: https://dentalkiosk.de
servers:
- url: https://cbshop.commerce-crew.com
  description: Production Middleware Server
- url: http://localhost:3330
  description: Local Development
...
paths:
...
  /api/cart/summary:
    post:
      operationId: getCartSummary
      summary: Get current cart summary
      description: 'Returns a simplified cart object (line item ids, quantities and
        totals) using the Shopware Store API.

        This endpoint is mainly used by the Shopware storefront widget to avoid brittle
        HTML parsing.

        '
      tags:
      - Cart
      parameters:
      - $ref: '#/components/parameters/ContextToken'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                context_token:
                  type: string
                  description: Optional context token (alternative to header)
      responses:
        '200':
          description: Cart summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartSummaryResponse'
  /api/cart/add:
    post:
      operationId: addToCart
      summary: Add product to shopping cart
      description: 'Queue an "add to cart" action for the currently active storefront
        user.

        The Shopware storefront widget will execute the action in the browser.


        Notes:

        - This is designed as a Dify tool endpoint.

        - It does NOT directly mutate the cart on the server.

        '
      tags:
      - Cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - product_id
              properties:
                product_id:
                  type: string
                  description: Shopware product UUID to add
                productId:
                  type: string
                  description: Alias for product_id
                product_name:
                  type: string
                  description: Optional product name (for better logs/UI)
                quantity:
                  type: integer
                  description: Quantity to add
                  default: 1
                  minimum: 1
                user_id:
                  type: string
                  description: Optional user id (conversation/userId from the widget)
      responses:
        '200':
          description: Action queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartToolActionResponse'
        '400':
          description: Validation error
  /api/cart/update:
    post:
      operationId: updateCartQuantity
      summary: Update quantity of a cart item
      description: 'Queue a "change quantity" action for the active storefront user.

        The storefront widget will resolve the correct line item id and execute the
        change.

        '
      tags:
      - Cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - quantity
              properties:
                line_item_id:
                  type: string
                  description: Preferred line item id (if known)
                lineItemId:
                  type: string
                  description: Alias for line_item_id
                product_id:
                  type: string
                  description: Product UUID (alternative to line_item_id)
                productId:
                  type: string
                  description: Alias for product_id
                product_name:
                  type: string
                  description: Optional product name (for better logs/UI)
                quantity:
                  type: integer
                  description: New quantity
                  minimum: 1
                  default: 1
                user_id:
                  type: string
                  description: Optional user id (conversation/userId from the widget)
      responses:
        '200':
          description: Action queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartToolActionResponse'
        '400':
          description: Validation error
  /api/cart/remove:
    post:
      operationId: removeFromCart
      summary: Remove an item from cart
      description: 'Queue a "remove item" action for the active storefront user.

        The storefront widget will execute the removal.

        '
      tags:
      - Cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                line_item_id:
                  type: string
                  description: Preferred line item id (if known)
                lineItemId:
                  type: string
                  description: Alias for line_item_id
                product_id:
                  type: string
                  description: Product UUID (alternative)
                productId:
                  type: string
                  description: Alias for product_id
                product_name:
                  type: string
                  description: Optional product name (for better logs/UI)
                user_id:
                  type: string
                  description: Optional user id (conversation/userId from the widget)
      responses:
        '200':
          description: Action queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartToolActionResponse'
        '400':
          description: Validation error
  /api/cart/clear:
    post:
      operationId: emptyCart
      summary: Empty the entire cart
      description: 'Queue a "clear cart" action for the active storefront user.

        The storefront widget will execute the removal.

        '
      tags:
      - Cart
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: string
                  description: Optional user id (conversation/userId from the widget)
      responses:
        '200':
          description: Action queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartToolActionResponse'
...
  /api/orders:
    post:
      operationId: getCustomerOrders
      summary: Get customer order history
      description: 'Retrieve past orders for a logged-in customer.

        Requires valid context token from logged-in session.

        '
      tags:
      - Customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                context_token:
                  type: string
                  description: Customer session token
                limit:
                  type: integer
                  default: 5
                  description: Maximum number of orders to return
      responses:
        '200':
          description: Order history
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  orders:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
                  count:
                    type: integer
                  error:
                    type: string
                    description: Error code if not logged in
                  message:
                    type: string
...
  /api/orders/last:
    post:
      operationId: getLastOrder
      summary: Get last order (newest)
      tags:
      - Orders
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
  /api/orders/by-number:
    post:
      operationId: getOrderByNumber
      summary: Get order by orderNumber
      tags:
      - Orders
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                orderNumber:
                  type: string
              required:
              - orderNumber
      responses:
        '200':
          description: OK
        '404':
          description: Not found
  /api/orders/reorder:
    post:
      operationId: reorder
      summary: Reorder an order (by orderNumber if provided)
      description: Enqueues cart add actions for each reorder item. The storefront
        widget executes them.
      tags:
      - Orders
      - Cart
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                orderNumber:
                  type: string
                  nullable: true
                clearFirst:
                  type: boolean
                  default: false
                userId:
                  type: string
      responses:
        '200':
          description: Queued actions
components:
  parameters:
    ContextToken:
      name: sw-context-token
      in: header
      required: false
      schema:
        type: string
      description: Customer session token for cart persistence
...
  schemas:
...
    CartSummaryResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        cart:
          type: object
          properties:
            items:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: string
                    description: Line item id
                  lineItemId:
                    type: string
                    description: Duplicate of id for clarity
                  productId:
                    type: string
                  name:
                    type: string
                  quantity:
                    type: integer
                  price:
                    type: number
                  totalPrice:
                    type: number
            itemCount:
              type: integer
            total:
              type: number
    CartToolActionResponse:
      type: object
      description: Response returned by tool endpoints that queue cart actions.
      properties:
        success:
          type: boolean
        message:
          type: string
        product_id:
          type: string
          nullable: true
        product_name:
          type: string
          nullable: true
        quantity:
          type: integer
          nullable: true
        line_item_id:
          type: string
          nullable: true
...
    Order:
      type: object
      properties:
        id:
          type: string
        orderNumber:
          type: string
        date:
          type: string
        dateFormatted:
          type: string
        total:
          type: number
        totalFormatted:
          type: string
        status:
          type: string
        items:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              name:
                type: string
              quantity:
                type: integer
              price:
                type: number
              priceFormatted:
                type: string